Renamed entity base classes
