+++ packages/core/src/ConnectorTypes.ts
@@ -1,7 +1,7 @@
-import App from './wrappers/App'
-import Permission from './wrappers/Permission';
-import Repo from './wrappers/Repo';
-import Role from './wrappers/Role';
+import App from './entities/App'
+import Permission from './entities/Permission';
+import Repo from './entities/Repo';
+import Role from './entities/Role';
 
 // TODO: No functions should be optional.
 // Made optional for now, so that we can develop one connector at a time.
