+++ packages/core/src/entities/App.ts
@@ -1,5 +1,5 @@
 import Repo from "./Repo"
-import Base from "./Base"
+import Entity from "./Entity"
 import { ConnectorInterface } from "../ConnectorTypes"
 
 // TODO: Implement all properties and methods from the API spec (https://github.com/aragon/plumbery/blob/master/docs/app.md).
@@ -39,7 +39,7 @@ export interface AppData {
   isForwarder: boolean
 }
 
-export default class App extends Base implements AppData {
+export default class App extends Entity implements AppData {
   readonly name?: string
   readonly address!: string
   readonly appId!: string
