+++ packages/core/src/entities/Repo.ts
@@ -1,4 +1,4 @@
-import Base from "./Base"
+import Entity from "./Entity"
 import { ConnectorInterface } from "../ConnectorTypes"
 
 // TODO: Implement all properties and methods from the API spec (https://github.com/aragon/plumbery/blob/master/docs/repo.md).
@@ -18,7 +18,7 @@ export interface RepoData {
   address: string
 }
 
-export default class Repo extends Base implements RepoData {
+export default class Repo extends Entity implements RepoData {
   readonly name!: string
   readonly address!: string
